frontend live

+++ routes/api.php
@@ -43,19 +43,18 @@ Route::prefix('v1')->group(function () {
 
         // **User Management**
         Route::get('user', [UserController::class, 'authUser']);
-        Route::get('users', [UserController::class, 'index'])->middleware('permission:View User');
-        Route::post('users', [UserController::class, 'store']);
-        Route::get('users/{user}', [UserController::class, 'show']);
-        Route::put('users/{user}', [UserController::class, 'update']);
-        Route::delete('users/{user}', [UserController::class, 'destroy']);
-
+        Route::get('users', [UserController::class, 'index'])->middleware('permission:view users');
+        Route::post('users', [UserController::class, 'store'])->middleware('permission:add users');
+        Route::get('users/{user}', [UserController::class, 'show'])->middleware('permission:view users');
+        Route::put('users/{user}', [UserController::class, 'update'])->middleware('permission:edit users');
+        Route::delete('users/{user}', [UserController::class, 'destroy'])->middleware('permission:delete users');
 
         // **Role Management**
-        Route::get('roles', [RoleController::class, 'index']);
-        Route::post('roles', [RoleController::class, 'store']);
-        Route::get('roles/{role}', [RoleController::class, 'show']);
-        Route::put('roles/{role}', [RoleController::class, 'update']);
-        Route::delete('roles/{role}', [RoleController::class, 'destroy']);
+        Route::get('roles', [RoleController::class, 'index'])->middleware('permission:view roles');
+        Route::post('roles', [RoleController::class, 'store'])->middleware('permission:add roles');
+        Route::get('roles/{role}', [RoleController::class, 'show'])->middleware('permission:view roles');
+        Route::put('roles/{role}', [RoleController::class, 'update'])->middleware('permission:edit roles');
+        Route::delete('roles/{role}', [RoleController::class, 'destroy'])->middleware('permission:delete roles');
 
 
         // **Task Management**
@@ -66,13 +65,13 @@ Route::prefix('v1')->group(function () {
         Route::delete('tasks/{id}', [TaskController::class, 'destroy']);
 
         
+        // **Task Management**
+        Route::get('employee', [EmployeeController::class, 'index']);
+        Route::post('employee', [EmployeeController::class, 'store']);
+        Route::get('employee/{id}', [EmployeeController::class, 'show']);
+        Route::put('employee/{id}', [EmployeeController::class, 'update']);
+        Route::delete('employee/{id}', [EmployeeController::class, 'destroy']);
         
     });
-    // **Task Management**
-    Route::get('employees', [EmployeeController::class, 'index']);
-    Route::post('employees', [EmployeeController::class, 'store']);
-    Route::get('employees/{id}', [EmployeeController::class, 'show']);
-    Route::put('employees/{id}', [EmployeeController::class, 'update']);
-    Route::delete('employees/{id}', [EmployeeController::class, 'destroy']);
     
 });
